refactoring: extract CartItem component from ShoppingCart
